Try to add support for intermediate CAs
It does not seem to work, though. Test with:

    import trustme

    ca = trustme.CA()
    ca = ca.create_child_ca()
    server_cert = ca.issue_server_cert(u"test-host.example.org")

    server_cert.private_key_and_cert_chain_pem.write_to_path("server.pem")

When commenting the `ca = ca.create_child_ca()` line, `openssl verify
-untrusted server.pem server.pem` says:

    server.pem: O = "trustme v0.4.0+dev", OU = Testing server cert #XukT8k24PBXQBSqv
    error 20 at 0 depth lookup:unable to get local issuer certificate

But with the `create_child_ca` line, we get:

    server.pem: O = "trustme v0.4.0+dev", OU = Testing CA #D2vixR6eI2g5N2k5
    error 19 at 1 depth lookup:self signed certificate in certificate chain

I have no idea what I'm doing: what's the problem? I did try to make
sure to only self-sign the root CA and not including it in the
server.pem file.