Allow tests to pass after 2038
in openSUSE, this broke tests in python-aiosmtplib,
python-anyio, python-cheroot, python-trio when running after 2038-01-01

Background:
As part of my work on reproducible builds for openSUSE, I check that software still gives identical build results in the future.
The usual offset is +16 years, because that is how long I expect some software will be used in some places.
This showed up failing tests in our package build.
See https://reproducible-builds.org/ for why this matters.